perf(detect): hold a Field's offset and length in int32
compile.Field was 32 bytes to carry two numbers that Compile refuses
over 255. Offset and Len were int because that is what the detectors
compute in, so each spent 8 bytes on 60 bits of headroom nothing could
reach. They are int32 now and a Field is 16 bytes, measured with
unsafe.Sizeof in a scratch test in internal/compile, deleted in the same
command.

Every fallback detector allocates a slice of these and they are the bulk
of what a non-trie parse allocates. Byte counts do not depend on the
machine, so these are exact:

    Parse_TextualMonth      608 B/op -> 448 B/op   -26.32%
    Parse_AmbiguousSlash    592 B/op -> 464 B/op   -21.62%
    Parse_ISOWeekDate       464 B/op -> 368 B/op   -20.69%
    Parse_ISOOrdinal        368 B/op -> 320 B/op   -13.04%
    Parse_ISO8601 (control) 208 B/op   unchanged
    VsStdlib (control)        0 B/op   unchanged

Allocation counts do not move: the slices are the same slices, each one
half the size. Parse_ISO8601 is flat because the trie formats carry
pre-built fields and allocate none of them, which is also why the warm
path cannot be affected in either direction: Program holds [24]Inst, Inst
is its own 6-byte type, and neither Program nor Layout changed. Program
is still 168 bytes and TestLayoutParseZeroAlloc is still exactly 0.

The nanoseconds are inconclusive and the run is not worth quoting: the
host sat at load average 18 on 12 cores throughout, every benchmark
including both controls came back p > 0.1, and VsStdlib, which cannot
reach any of this, "moved" 60%. Less zeroing and a smaller cache
footprint should help slightly; nothing here measured it.

The risk in a change like this is a truncating cast, which produces a
field whose declared offset is wrong the same way every time it runs.
That is C9's failure shape and comparing Parse against Layout.Parse
cannot see it. Two things close it.

The conversions are at the boundary. Detectors keep computing in int and
convert where a Field is built; the two loops that read Field.Offset back
convert once into int locals rather than doing int32 arithmetic. So there
is no mixed-width arithmetic anywhere, only a width change at the struct.

And Detect now refuses an input longer than math.MaxInt32 instead of
letting a position wrap into a small positive offset that looks valid.
Detect is the only entry point to the package, so one check covers every
detector. No input that long can have a correct answer anyway: Compile
refuses any field past byte 255, so every successful parse describes
bytes well inside the first 256.

What catches a mistake if both of those fail:
TestEveryInputByteIsDescribedExactlyOnce, which marks bytes rather than
summing widths, and the FixedWidth cross-check, which compares each
field's declared width against what its op reads. Both exist for this
class. The full suite runs green with -race.

+++ internal/compile/compile_test.go
@@ -284,7 +284,7 @@ func TestCompileRefusesWhatItCannotAddress(t *testing.T) {
 func TestCompileRefusesTooManyInstructions(t *testing.T) {
 	fields := make([]Field, MaxInstructions+1)
 	for i := range fields {
-		fields[i] = Field{Kind: FLiteral, Offset: i, Len: 1, Aux: 'x'}
+		fields[i] = Field{Kind: FLiteral, Offset: int32(i), Len: 1, Aux: 'x'}
 	}
 	if _, _, err := Compile(&FormatDef{Name: "PROBE", Fields: fields}, time.UTC); err == nil {
 		t.Errorf("Compile accepted %d fields, the limit is %d", len(fields), MaxInstructions)
@@ -401,8 +401,8 @@ func compileUnfused(def *FormatDef, tz *time.Location) Program {
 // separator was its own instruction with its own refusal, and folding it must
 // not turn it into a byte nobody looks at.
 func TestFusionMatchesTheUnfusedProgram(t *testing.T) {
-	classLit := func(off int) Field { return Field{Kind: FLiteral, Offset: off, Len: 1} }
-	exactLit := func(off int, b byte) Field {
+	classLit := func(off int32) Field { return Field{Kind: FLiteral, Offset: off, Len: 1} }
+	exactLit := func(off int32, b byte) Field {
 		return Field{Kind: FLiteral, Offset: off, Len: 1, Aux: uint16(b)}
 	}
